requirements.txt
asgiref==3.5.2
autopep8==1.7.0
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.0.8
django-cors-headers==3.13.0
djangorestframework==3.14.0
idna==3.4
pycodestyle==2.9.1
pytz==2022.4
PyYAML==6.0
requests==2.28.1
sqlparse==0.4.3
toml==0.10.2
tzdata==2022.4
urllib3==1.26.12
